+++ cedar-policy-core/src/tpe/response.rs
@@ -181,33 +181,57 @@ impl<'a> Response<'a> {
     }
 
     /// Get policy ids of satisified permit residual policies
-    pub fn satisfied_permits(&self) -> impl Iterator<Item = &PolicyID> {
-        self.satisfied_permits.iter()
+    pub fn satisfied_permits(&self) -> impl Iterator<Item = &Policy> {
+        // PANIC SAFETY: we know that the policy ids are in the residuals map
+        #[allow(clippy::unwrap_used)]
+        self.satisfied_permits
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap().policy.as_ref())
     }
 
     /// Get policy ids of satisified forbid residual policies
-    pub fn satisfied_forbids(&self) -> impl Iterator<Item = &PolicyID> {
-        self.satisfied_forbids.iter()
+    pub fn satisfied_forbids(&self) -> impl Iterator<Item = &Policy> {
+        // PANIC SAFETY: we know that the policy ids are in the residuals map
+        #[allow(clippy::unwrap_used)]
+        self.satisfied_forbids
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap().policy.as_ref())
     }
 
     /// Get policy ids of trivially false permit residual policies
-    pub fn false_permits(&self) -> impl Iterator<Item = &PolicyID> {
-        self.false_permits.iter()
+    pub fn false_permits(&self) -> impl Iterator<Item = &Policy> {
+        // PANIC SAFETY: we know that the policy ids are in the residuals map
+        #[allow(clippy::unwrap_used)]
+        self.false_permits
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap().policy.as_ref())
     }
 
     /// Get policy ids of trivially false forbid residual policies
-    pub fn false_forbids(&self) -> impl Iterator<Item = &PolicyID> {
-        self.false_forbids.iter()
+    pub fn false_forbids(&self) -> impl Iterator<Item = &Policy> {
+        // PANIC SAFETY: we know that the policy ids are in the residuals map
+        #[allow(clippy::unwrap_used)]
+        self.false_forbids
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap().policy.as_ref())
     }
 
     /// Get policy ids of non-trivial permit residual policies
-    pub fn residual_permits(&self) -> impl Iterator<Item = &PolicyID> {
-        self.residual_permits.iter()
+    pub fn residual_permits(&self) -> impl Iterator<Item = &Policy> {
+        // PANIC SAFETY: we know that the policy ids are in the residuals map
+        #[allow(clippy::unwrap_used)]
+        self.residual_permits
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap().policy.as_ref())
     }
 
     /// Get policy ids of non-trivial forbid residual policies
-    pub fn residual_forbids(&self) -> impl Iterator<Item = &PolicyID> {
-        self.residual_forbids.iter()
+    pub fn residual_forbids(&self) -> impl Iterator<Item = &Policy> {
+        // PANIC SAFETY: we know that the policy ids are in the residuals map
+        #[allow(clippy::unwrap_used)]
+        self.residual_forbids
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap().policy.as_ref())
     }
 
     /// Look up the [`Residual`] by [`PolicyID`]
@@ -248,10 +272,7 @@ impl<'a> Response<'a> {
     }
 
     /// Get residual policies
-    pub fn residual_policies(&self) -> Vec<Policy> {
-        self.residuals
-            .values()
-            .map(|rp| rp.clone().into())
-            .collect()
+    pub fn residual_policies(&self) -> impl Iterator<Item = &Policy> {
+        self.residuals.values().map(|rp| rp.policy.as_ref())
     }
 }