+++ cedar-policy/src/api.rs
@@ -5331,16 +5331,16 @@ mod tpe {
     /// A partial version of [`crate::Response`].
     #[repr(transparent)]
     #[derive(Debug, Clone, RefCast)]
-    pub struct TPEResponse<'a>(pub(crate) tpe::response::Response<'a>);
+    pub struct TpeResponse<'a>(pub(crate) tpe::response::Response<'a>);
 
     #[doc(hidden)]
-    impl<'a> AsRef<tpe::response::Response<'a>> for TPEResponse<'a> {
+    impl<'a> AsRef<tpe::response::Response<'a>> for TpeResponse<'a> {
         fn as_ref(&self) -> &tpe::response::Response<'a> {
             &self.0
         }
     }
 
-    impl TPEResponse<'_> {
+    impl TpeResponse<'_> {
         /// Attempt to get the authorization decision
         pub fn decision(&self) -> Option<Decision> {
             self.0.decision()
@@ -5356,6 +5356,20 @@ mod tpe {
                 .reauthorize(&request.0, &entities.0)
                 .map(Into::into)
                 .map_err(Into::into)
+        }
+
+        /// Return residual policies for each policy in the input policy set
+        /// Use [`TpeResponse::nontrivial_residual_policies`] to get non-trivial residual policies
+        pub fn residual_policies(&self) -> impl Iterator<Item = Policy> + '_ {
+            self.0.residual_policies().map(|p| p.clone().into())
+        }
+
+        /// Returns an iterator of non-trivial (meaning more than just `true` or `false`) residual policies
+        pub fn nontrivial_residual_policies(&'_ self) -> impl Iterator<Item = Policy> + '_ {
+            self.0
+                .residual_permits()
+                .chain(self.0.residual_forbids())
+                .map(|p| p.clone().into())
         }
     }
 
@@ -5432,11 +5446,11 @@ mod tpe {
             request: &'a PartialRequest,
             entities: &'a PartialEntities,
             schema: &'a Schema,
-        ) -> Result<TPEResponse<'a>, tpe_err::TPEError> {
+        ) -> Result<TpeResponse<'a>, tpe_err::TPEError> {
             use cedar_policy_core::tpe::is_authorized;
             let ps = &self.ast;
             let res = is_authorized(ps, &request.0, &entities.0, &schema.0)?;
-            Ok(TPEResponse(res))
+            Ok(TpeResponse(res))
         }
 
         /// Like [`Authorizer::is_authorized`] but uses an [`EntityLoader`] to load
@@ -5479,7 +5493,7 @@ mod tpe {
                     .0
                     .residual_policies()
                     .into_iter()
-                    .map(Policy::from_ast),
+                    .map(|p| p.clone().into()),
             )
             .unwrap();
             // PANIC SAFETY: request construction should succeed because each entity passes validation
@@ -5532,7 +5546,7 @@ mod tpe {
                     .0
                     .residual_policies()
                     .into_iter()
-                    .map(Policy::from_ast),
+                    .map(|p| p.clone().into()),
             )
             .unwrap();
             // PANIC SAFETY: request construction should succeed because each entity passes validation